🚀 Refactor: Mejoras en la gestión de datos de usuario y optimización de efectos 🔹 Refactorizado InfoUser.jsx para mejorar la obtención de datos desde localStorage 🔹 Implementado efecto con setInterval para actualizar datos en tiempo real 🔹 Agregado console.log para monitorear cambios en localStorage 🔹 Creado componente Tasklist_compoment para mejorar modularidad en la visualización de tareas 🔹 Optimizada InfoUser_component.jsx para mostrar datos completos del usuario
